eng_gls_backend/requirements.txt
asgiref==3.4.1
astor==0.8.1
attrs==21.2.0
bandit==1.7.0
darglint==1.8.0
Django==3.2.8
django-cors-headers==3.10.0
django-environ==0.8.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
docutils==0.17.1
eradicate==2.0.0
flake8==3.9.2
flake8-bandit==2.1.2
flake8-broken-line==0.3.0
flake8-bugbear==21.9.2
flake8-commas==2.0.0
flake8-comprehensions==3.7.0
flake8-debugger==4.0.0
flake8-docstrings==1.6.0
flake8-eradicate==1.1.0
flake8-isort==4.0.0
flake8-polyfill==1.0.2
flake8-quotes==3.3.0
flake8-rst-docstrings==0.2.3
flake8-string-format==0.3.0
gitdb==4.0.7
GitPython==3.1.24
gunicorn==20.1.0
isort==5.9.3
mccabe==0.6.1
pbr==5.6.0
pep8-naming==0.11.1
psycopg2==2.9.1
pycodestyle==2.7.0
pydocstyle==6.1.1
pyflakes==2.3.1
Pygments==2.10.0
PyJWT==2.2.0
pytz==2021.3
PyYAML==5.4.1
restructuredtext-lint==1.3.2
six==1.16.0
smmap==4.0.0
snowballstemmer==2.1.0
sqlparse==0.4.2
stevedore==3.4.0
testfixtures==6.18.3
typing-extensions==3.10.0.2
wemake-python-styleguide==0.15.3
